+++ nisystemlink/clients/testmonitor/utilities/__init__.py
@@ -1,3 +1,3 @@
-from ._dataframe_utility import get_results_dataframe
+from ._dataframe_utilities import get_results_dataframe
 
 # flake8: noqa
