+++ nisystemlink/clients/testmonitor/utilities/_constants.py
@@ -1,0 +1,1 @@
+DEFAULT_QUERY_RESULTS_TAKE = 1000
