+++ nisystemlink/clients/testmonitor/utilities/_dataframe_utilities.py
@@ -1,0 +1,50 @@
+from typing import List, Optional
+
+import pandas as pd
+from nisystemlink.clients.testmonitor._test_monitor_client import TestMonitorClient
+from nisystemlink.clients.testmonitor.models import (
+    Result,
+    ResultProjection,
+)
+from nisystemlink.clients.testmonitor.utilities._client_utilities import (
+    __batch_query_results,
+)
+
+
+def get_results_dataframe(
+    client: TestMonitorClient,
+    query_filter: str,
+    column_projection: Optional[List[ResultProjection]] = None,
+) -> pd.DataFrame:
+    """Fetches results of a specific product and normalizes them into a Pandas DataFrame.
+
+    Args:
+        client: The TestMonitorClient to fetch results data.
+        query_filter: The filter to use when querying the results.
+        column_projection: List of columns to retrieve when querying the results.
+            Fields you do not specify are excluded. Returns all fields if no value is specified.
+
+    Returns:
+        A Pandas DataFrame containing the results.
+    """
+    queried_results = __batch_query_results(client, query_filter, column_projection)
+
+    results_dataframe = __normalize_results(queried_results)
+
+    return results_dataframe
+
+
+def __normalize_results(results: List[Result]) -> pd.DataFrame:
+    """Normalizes the results into a Pandas DataFrame.
+
+    Args:
+        results: The list of results to normalize.
+
+    Returns:
+        A Pandas DataFrame with the normalized queried results.
+    """
+    results_dict = [results.dict(exclude_unset=True) for results in results]
+    normalized_dataframe = pd.json_normalize(results_dict, sep=".")
+    normalized_dataframe.dropna(axis="columns", how="all", inplace=True)
+
+    return normalized_dataframe
